backend/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
black==24.10.0
boto3==1.35.98
botocore==1.35.98
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email-validator==2.2.0
# emergentintegrations==0.1.0  # Not available on PyPI - removed
fastapi==0.110.1
fastuuid==0.14.0
filelock==3.13.1
flake8==7.1.1
frozenlist==1.5.0
fsspec==2024.12.0
# google-ai-generativelanguage==0.6.11
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-genai==0.2.2
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
hf-xet==0.1.0
httpcore==1.0.7
httplib2==0.22.0
# httpx==0.28.1
huggingface_hub==0.27.1
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
jq==1.8.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.57.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
motor==3.3.1
multidict==6.1.0
mypy==1.14.1
mypy_extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
openai
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.5.0
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytokens==0.3.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
s3transfer==0.10.4
s5cmd==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.37.2
# stripe==11.4.0
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
typer==0.15.1
typer-slim==0.15.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==14.1
yarl==1.18.3
zipp==3.21.0
